TASK-101: address review follow-ups
