+++ lib/services/project-attachment-service.ts
@@ -421,29 +421,25 @@ export async function createTaskAttachmentFromForm(input: {
       }
 
       try {
-        const attachment = await db.$transaction(async (tx) => {
-          const createdAttachment = await tx.taskAttachment.create({
-            data: {
-              taskId: input.taskId,
-              uploadedByUserId: actorUserId,
-              kind: ATTACHMENT_KIND_LINK,
-              name: providedName || new URL(normalizedUrl).hostname,
-              url: normalizedUrl,
-            },
-            select: {
-              id: true,
-              kind: true,
-              name: true,
-              url: true,
-              mimeType: true,
-              sizeBytes: true,
-            },
-          });
-
-          await touchTaskActivity(tx, input.taskId, actorUserId);
-
-          return createdAttachment;
+        const attachment = await db.taskAttachment.create({
+          data: {
+            taskId: input.taskId,
+            uploadedByUserId: actorUserId,
+            kind: ATTACHMENT_KIND_LINK,
+            name: providedName || new URL(normalizedUrl).hostname,
+            url: normalizedUrl,
+          },
+          select: {
+            id: true,
+            kind: true,
+            name: true,
+            url: true,
+            mimeType: true,
+            sizeBytes: true,
+          },
         });
+
+        await touchTaskActivity(db, input.taskId, actorUserId);
 
         return {
           ok: true,
@@ -497,31 +493,27 @@ export async function createTaskAttachmentFromForm(input: {
         resolveAttachmentMimeType(storedFile.mimeType, storedFile.originalName) ??
         resolvedFormMimeType;
 
-      const attachment = await db.$transaction(async (tx) => {
-        const createdAttachment = await tx.taskAttachment.create({
-          data: {
-            taskId: input.taskId,
-            uploadedByUserId: actorUserId,
-            kind: ATTACHMENT_KIND_FILE,
-            name: providedName || storedFile.originalName,
-            storageKey: storedFile.storageKey,
-            mimeType: resolvedStoredMimeType,
-            sizeBytes: storedFile.sizeBytes,
-          },
-          select: {
-            id: true,
-            kind: true,
-            name: true,
-            url: true,
-            mimeType: true,
-            sizeBytes: true,
-          },
-        });
-
-        await touchTaskActivity(tx, input.taskId, actorUserId);
-
-        return createdAttachment;
+      const attachment = await db.taskAttachment.create({
+        data: {
+          taskId: input.taskId,
+          uploadedByUserId: actorUserId,
+          kind: ATTACHMENT_KIND_FILE,
+          name: providedName || storedFile.originalName,
+          storageKey: storedFile.storageKey,
+          mimeType: resolvedStoredMimeType,
+          sizeBytes: storedFile.sizeBytes,
+        },
+        select: {
+          id: true,
+          kind: true,
+          name: true,
+          url: true,
+          mimeType: true,
+          sizeBytes: true,
+        },
       });
+
+      await touchTaskActivity(db, input.taskId, actorUserId);
 
       return {
         ok: true,
@@ -895,31 +887,27 @@ export async function finalizeTaskAttachmentDirectUpload(input: {
         );
       }
 
-      const attachment = await db.$transaction(async (tx) => {
-        const createdAttachment = await tx.taskAttachment.create({
-          data: {
-            taskId: input.taskId,
-            uploadedByUserId: actorUserId,
-            kind: ATTACHMENT_KIND_FILE,
-            name: normalizedUpload.data.name,
-            storageKey: normalizedStorageKey,
-            mimeType: resolvedMimeType,
-            sizeBytes: resolvedSizeBytes,
-          },
-          select: {
-            id: true,
-            kind: true,
-            name: true,
-            url: true,
-            mimeType: true,
-            sizeBytes: true,
-          },
-        });
-
-        await touchTaskActivity(tx, input.taskId, actorUserId);
-
-        return createdAttachment;
+      const attachment = await db.taskAttachment.create({
+        data: {
+          taskId: input.taskId,
+          uploadedByUserId: actorUserId,
+          kind: ATTACHMENT_KIND_FILE,
+          name: normalizedUpload.data.name,
+          storageKey: normalizedStorageKey,
+          mimeType: resolvedMimeType,
+          sizeBytes: resolvedSizeBytes,
+        },
+        select: {
+          id: true,
+          kind: true,
+          name: true,
+          url: true,
+          mimeType: true,
+          sizeBytes: true,
+        },
       });
+
+      await touchTaskActivity(db, input.taskId, actorUserId);
 
       return {
         ok: true,
@@ -1385,13 +1373,11 @@ export async function deleteTaskAttachmentForProject(input: {
         return createError(404, "Attachment not found");
       }
 
-      await db.$transaction(async (tx) => {
-        await tx.taskAttachment.delete({
-          where: { id: attachment.id },
-        });
-
-        await touchTaskActivity(tx, input.taskId, actorUserId);
+      await db.taskAttachment.delete({
+        where: { id: attachment.id },
       });
+
+      await touchTaskActivity(db, input.taskId, actorUserId);
 
       if (attachment.kind === ATTACHMENT_KIND_FILE && attachment.storageKey) {
         await deleteAttachmentFile(attachment.storageKey).catch((error) => {
